fix(autocomplete, filter-field): Fixes an issue with overlay positioning.
BREAKING CHANGE: Since the constructor signatures of the autocomplete
and the filter-field-range have changed, this is considered a breaking
change which will be released with 6.0.0

Switched the autocomplete and filterfield overlay to use the
DtFlexibleConnectedPositionStrategy instead of the CDK
FlexibleConnectedPositionStrategy. The Dt version can deal with the
clusters page-ribbon component and in general deals with better with
overlaying elements.

Fixes #451